#8 refactor(banner): revise banner module code struct
