+++ src/banner/util.rs
@@ -16,17 +16,6 @@ pub fn label(s: &str) -> String {
     format!("{}{}", s, " ".repeat(padding))
 }
 
-pub fn human_size(bytes: u64) -> String {
-    const UNITS: [&str; 4] = ["B", "KB", "MB", "GB"];
-    let mut size = bytes as f64;
-    let mut idx = 0;
-    while size >= 1024.0 && idx < UNITS.len() - 1 {
-        size /= 1024.0;
-        idx += 1;
-    }
-    format!("{:.2} {}", size, UNITS[idx])
-}
-
 #[cfg(windows)]
 pub fn canon(path: &Path) -> PathBuf {
     canonicalize(path).unwrap_or(path.to_path_buf())
